Refactored ForumThread DAO, Service and Controller
	- Changed update URL;
	- Added NotFound responses;
	- Added checks to see if data provided is correct or not.

+++ src/main/java/thecerealkillers/elearning/service/impl/ForumThreadServiceImpl.java
@@ -1,6 +1,7 @@
 package thecerealkillers.elearning.service.impl;
 
 
+import thecerealkillers.elearning.exceptions.NotFoundException;
 import thecerealkillers.elearning.exceptions.ServiceException;
 import thecerealkillers.elearning.service.ForumThreadService;
 import thecerealkillers.elearning.exceptions.DAOException;
@@ -8,6 +9,8 @@ import thecerealkillers.elearning.dao.ForumThreadDAO;
 import thecerealkillers.elearning.model.ForumThread;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import thecerealkillers.elearning.service.TopicService;
+import thecerealkillers.elearning.service.UserService;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -22,17 +25,36 @@ public class ForumThreadServiceImpl implements ForumThreadService {
     @Autowired
     private ForumThreadDAO forumThreadDAO;
 
+    @Autowired
+    private UserService userService;
+
+    @Autowired
+    private TopicService topicService;
+
+
     @Override
-    public void add(ForumThread newForumThread, String topicTitle) throws ServiceException {
+    public void add(ForumThread newForumThread, String topicTitle) throws ServiceException, NotFoundException {
         try {
-            forumThreadDAO.add(newForumThread, topicTitle);
+            if (!exists(newForumThread.getTitle())) {
+                if (userService.usernameExists(newForumThread.getOwner())) {
+                    if (topicService.exists(topicTitle)) {
+                        forumThreadDAO.add(newForumThread, topicTitle);
+                    } else {
+                        throw new NotFoundException(NotFoundException.NO_TOPIC);
+                    }
+                } else {
+                    throw new NotFoundException(NotFoundException.NO_USER);
+                }
+            } else {
+                throw new ServiceException(ServiceException.THREAD_ALREADY_EXISTS);
+            }
         } catch (DAOException daoException) {
             throw new ServiceException(ServiceException.ADD_FORUM_THREAD);
         }
     }
 
     @Override
-    public List<ForumThread> getAll() throws ServiceException {
+    public List<ForumThread> getAll() throws ServiceException, NotFoundException {
         try {
             return forumThreadDAO.getAll();
         } catch (DAOException daoException) {
@@ -41,47 +63,88 @@ public class ForumThreadServiceImpl implements ForumThreadService {
     }
 
     @Override
-    public List<ForumThread> getThreadsOwnedByUser(String userName) throws ServiceException {
+    public List<ForumThread> getThreadsOwnedByUser(String userName) throws ServiceException, NotFoundException {
         try {
-            return forumThreadDAO.getThreadsOwnedByUser(userName);
+            if (userService.usernameExists(userName)) {
+                return forumThreadDAO.getThreadsOwnedByUser(userName);
+            } else {
+                throw new NotFoundException(NotFoundException.NO_USER);
+            }
         } catch (DAOException daoException) {
             throw new ServiceException(ServiceException.GET_THREADS_BY_OWNER);
         }
     }
 
     @Override
-    public ForumThread getThreadByTitle(String threadTitle) throws ServiceException {
+    public ForumThread getThreadByTitle(String threadTitle) throws ServiceException, NotFoundException {
         try {
-            return forumThreadDAO.getThreadByTitle(threadTitle);
+            if (exists(threadTitle)) {
+                return forumThreadDAO.getThreadByTitle(threadTitle);
+            } else {
+                throw new NotFoundException(NotFoundException.NO_THREAD);
+            }
         } catch (DAOException daoException) {
             throw new ServiceException(ServiceException.GET_THREAD_BY_TITLE);
         }
     }
 
     @Override
-    public List<ForumThread> getThreadsForTopic(String topic) throws ServiceException {
+    public List<ForumThread> getThreadsForTopic(String topic) throws ServiceException, NotFoundException {
         try {
-            return forumThreadDAO.getThreadsForTopic(topic);
+            if (topicService.exists(topic)) {
+                return forumThreadDAO.getThreadsForTopic(topic);
+            } else {
+                throw new NotFoundException(NotFoundException.NO_TOPIC);
+            }
         } catch (DAOException daoException) {
             throw new ServiceException(ServiceException.GET_THREADS_TOPIC);
         }
     }
 
     @Override
-    public void updateThread(String oldTitle, ForumThread newThread) throws ServiceException {
+    public void updateThread(String oldTitle, ForumThread newThread) throws ServiceException, NotFoundException {
         try {
-            forumThreadDAO.updateThread(oldTitle, newThread);
+            if (exists(oldTitle)) {
+                if (!exists(newThread.getTitle())) {
+                    if (userService.usernameExists(newThread.getOwner())) {
+                        forumThreadDAO.updateThread(oldTitle, newThread);
+                    } else {
+                        throw new NotFoundException(NotFoundException.NO_USER);
+                    }
+                } else {
+                    throw new ServiceException(ServiceException.THREAD_ALREADY_EXISTS);
+                }
+            } else {
+                throw new NotFoundException(NotFoundException.NO_THREAD);
+            }
         } catch (DAOException daoException) {
             throw new ServiceException(ServiceException.UPDATE_THREAD);
         }
     }
 
     @Override
-    public void deleteThreadByTitle(String threadToDelete) throws ServiceException {
+    public void deleteThreadByTitle(String threadToDelete) throws ServiceException, NotFoundException {
         try {
-            forumThreadDAO.deleteThreadByTitle(threadToDelete);
+            if (exists(threadToDelete)) {
+                forumThreadDAO.deleteThreadByTitle(threadToDelete);
+            } else {
+                throw new NotFoundException(NotFoundException.NO_THREAD);
+            }
         } catch (DAOException daoException) {
             throw new ServiceException(ServiceException.DELETE_THREAD_BY_TITLE);
+        }
+    }
+
+    @Override
+    public Boolean exists(String title) throws ServiceException {
+        try {
+            forumThreadDAO.getThreadByTitle(title);
+
+            return true;
+        } catch (NotFoundException e) {
+            return false;
+        } catch (DAOException e) {
+            throw new ServiceException(ServiceException.GET_THREAD_BY_TITLE);
         }
     }
 }
